Add files via upload
I worked on various ML Models for predicting CO2 Emissions with data sourced from Kaggle. The aim was to analyze and find the best model to accurately forecast CO2 emissions based on various features.

CatBoost Regressor turned out to be a game-changer in my analysis💡!
This model outperformed others, indicated by its relatively low RMSE (Root Mean Square Error) of 0.3643. Considering the wide range of CO2 emissions in the dataset (from approximately 0.000078 to 46.2635), this level of accuracy is a testament to the model's robustness.
This project enhanced my skills in applying machine learning models and underscored the importance of choosing the right model based on the data characteristics and the target variable. The CatBoost Regressor, with its ability to handle categorical data efficiently, proved to be the most suitable for this dataset.